split cmgen into an exe + libibl.a
