+++ libs/ibl/include/ibl/Image.h
